libsystemd: Make basic types accessors typesafe
This pulls the basic types logic out of the highlevel code.

Annoyingly, due to a minor inconsistency in the systemd API, we need two
functions.  On the way out, the correct type for 's' is 'char **', as
expected, but on the way in, it's 'char *'.  That's a nice convenience
for C users, but it means we need to work a bit harder to avoid byref()
for the "in" case.

There's also some kind of a start to a convention for low-level and
high-level librarywrapper APIs: all type and memory safety should be
achieved in the low-level code.  It should not be possible to leak or
crash due to a bug in the high-level code (assuming it's not "evil").