librarywrapper: use a wrapper for instancemethod
Unsurprisingly, Python's 'function' type implements the descriptor
protocol more efficiently than we can.  Using a wrapper here speeds
things up by ~10% across the board.

We're also able to remove the special-case we added for the
argument-list mangling of instance methods and move it to the wrapper,
cleaning up the method registration code a bit.