bus: tweak peek_type() invocation
We call this function an awful lot, so try to speed up our usage of it:
drop the extra wrapper and super() call and try to reuse the ctypes
boxes a bit.

This is a 17% improvement in Message.get_body() for a moderate-sized
Properties.GetAll (systemd-hostnamed) call and more should be possible.