bus: raise exceptions on DBus errors during calls
Export a new BusError exception type for errors during calls.  Throw it
from .call() and .call_async().

Properly export BusMessage while we're at it.

A note: the sd_bus_error field of the message handler function is called
`ret_error` and is used for reporting errors from the callback to
sd-bus, not the other way around.  For the async call variant, the
documentation specifically mentions that you need to inspect the
message.