Add initial binding for slots
Add a new high-level type called Slot (and make PendingCall a subclass
of it).

We use a bit of a trick here: we always initialise the high-level parts
of the Slot object before the call, but the pointer value will be NULL
at first.  When the call happens, it fills in the pointer, which
contains more or less the same information as we already have in the
object.

There's one very tricky thing to note: the Slot object holds the
callback.  That's necessary: the callback object is a ctypes-computed
trampoline (and a closure, at that) and we need to make sure it stays
alive until C can call it.  Hanging this off of the same object which
controls the lifecycle of the callback and userdata on the "C side" is a
rather nice fit.  This will break down if someone else takes a ref on
the slot object, but nobody ought to be doing that.  We might imagine a
more robust solution where we use the destroy notify mechanism of the
slot to free the callback trampoline, but doing this in a reasonable way
(ie: one that doesn't involve keeping the callback alive in some kind of
a global list) is sure to be thorny.

There should probably also be an explicit mechanism to unref the slot
without relying on garbage collection.