workflows: simplify dependencies, use py3 tox env
The default env in the tox config is `site`, which uses the
site-installed packages.  That's mostly intended for local use (to avoid
downloading a bunch of pip packages onto my laptop).

For CI, though, let's simplify our dependency situation by using an
isolated environment.  systemd_ctypes has no python dependencies itself:
it's just the testing framework bits that we want to install.

Note: we want to ensure that dbus-python is installed on the host, since
we don't want to have to compile it ourselves inside the venv.  Bump the
`sitepackages = True` line to all environments in the tox config,
accordingly.  We install dbusmock on the host, even though we don't need
it, since it pulls in the dependency, plus some necessary others (dbus
libraries, dbus-daemon, etc).