bus: add and test manual message construction
Add a new method for constructing a message one argument at a time.
This is what we're going to want to use from Cockpit.

Write a test that constructs a message in both ways and checks it.  We
need to bind sd_bus_message_seal() for that.  This test also exercises
encoding and decoding of structure fields.

The signature code is now fully covered.