bus: sd_bus_error handling cleanup
sd_bus_error is a pretty tricky type to bind.  Normally, this type is
meant to be stack-allocated, which is not an option from Python.  We
need to take a much more explicit approach to memory management, which
complicates things: we need to dynamically allocate a memory region,
which systemd allocates the contents of, and then (and only in the case
that we allocated it) we need to have systemd free its contents, before
we free it.

We must also never write to the structure: it's systemd's job to fill it
in.

Our current approach was broken in two ways:

 - it didn't allow us to set error messages

 - it leaked the error every time we received one as a result of a
   method call

We try a different approach which allows setting the error message (via
systemd's API) and we also conditionally request systemd to clear the
structure, but only in the case that we are responsible for having
allocated it.

We also add some API to make it easier to send error replies, without
dealing with sd_bus_error at all.