bus: add a new new() method
This method takes a number of keyword arguments and should act as a
single method for setting up all kinds of bus objects as either clients
or servers, and from fds or paths.

We also change the default attach_event value to True.  This is what you
want, most of the time.

Remove the incorrect duplicate low-level binding which stated that new()
was an instance method.

meganew