bus: up our object-exporting game
Add a series of new helper types to make it nicer to export objects on
the bus.  There is now signals support, and simplified support for
properties (not requiring getters or setters).

The interface decorator has been removed and replaced with a new
approach: each interface should be defined in its own subclass, and the
interface name is derived from the subclass name (with support for
overrides via kwarg=).

We add the standard Introspectable, Peer, and Properties interfaces via
the new mechanism and now define our 'Object' type as a BaseObject which
additionally provides all of those interfaces, as well as giving the
user a base to define their own interface (or interfaces).

Documentation has been expanded, and examples added.

There's also a .registered() method to notify an object when it joins
the bus.