bus: Fix return values from coroutines
If the handler function for a method call was a coroutine, we were
passing its value directly to .reply_method_return(), which expected to
have the return values each as a separate argument.  This worked fine if
we were returning one value, but broke for the `None` and multiple-value
cases.

Pass the return value back through .reply_method_function_return_value()

Add some more test cases with single, multiple and no return values for
async and non-async functions.