bus: Avoid reference loops in Slot and PendingCall
Instead of taking an indirect references on ourselves via bound method
calls, use closures in the form of nested functions.

This means that the Slot can be immediately freed on unref, without
waiting for the GC to run.  That's important, since this can result in
the bus being held open, otherwise.

All the same, let's add an explicit .cancel() method there as another
way of releasing the slot.  This shouldn't be required.

Add a previously failing test-case for unexporting.

Note: we handle the reference to the closure completely within the
Python realm, and C has absolutely no reference to anything — just a
dangling pointer.  We know that when the Slot goes out of scope, we'll
unref the slot in C, which will prevent further callbacks, so this
should be safe.  The reason we want to keep all references inside of
Python (instead of using the somewhat safer destroy notify mechanism
from C, for example) is to allow the GC to detect and bust cycles.  We
add a somewhat exotic testcase to verify that the GC (and only the GC)
is capable of busting exactly the cycles, in a way similar to how this
will be used from Cockpit.