Handle: don't warn on implicit close
Follow the behaviour of other Python file-like objects and simply close
our fd when we go out of scope, instead of failing an assert.

Keeping track of everything and making sure it gets closed isn't always
reasonable: particularly at program exit, and particularly when the
program exists due to an exception.