bus: replace type system
Add a new "bustypes" module, replacing the previous "signature" module.
This implements a significantly more formalized, robust and performant
approach to (de)serializing data, centralized in a single location,
instead of being spread out across three separate files.

"Formalized": each dbus type is represented by a Type object which has
reader and writer functions for that type.  The complex types call the
reader/writer functions of their contained types.

"Robust": we go to great efforts to detect and report errors, and
include many unit tests to this effect.

"Performant": depending on the type, informal benchmarks show a round-trip
through (de)serializing should be on the order of 3-80x faster.  The 80x
is for moderate-sized instances of the 'ay' type, which is now sent as a
byte array (rather than as one integer as a time).

Some extra improvements:

 - support for specifying Python types instead of D-Bus typestrings, in
   many places.  A new BusType enum is defined with constants like
   uint16, byte, objectpath, etc.

 - binary data is now handled primarily via the 'bytes' type, although
   memoryview and bytearray are also supported as input.  Base64-encoded
   strings can also be transparently accepted as input, but won't be
   output.  Instead, we now define a small JSONEncoder subclass which
   knows how to encode bytes instances as base64 strings.

 - similarly, we introduce a new Variant type to replace the dictionary
   based {"t" / "v"} format that we used before.  The old format is
   still accepted as input, but won't be output.  Our JSONEncoder knows
   how to encode Variant to the dict-based format.