bus: add support for OSError exceptions
Allow passing OSError to various 'error reply' functions, as well as
catching it and forwarding it from method calls exported on the bus.

The error will be converted to the usual org.freedesktop.DBus.Error
message based on its errno and the str() form of the exception will be
used as its message.