Replace C binding layer with static typing hints
Instead of our own custom datatype for representing the functions that
we want to bind, along with their type information, introduce a new
system based on static typing hints.

In the new system, we define the classes and methods as if they were
to be implemented in Python, but we exclude method bodies.  This allows
static checking to reason able things correctly.  The binding layer then
goes through the classes and replaces the methods with the C function
pointers, using the static type hints to decide which C types to use for
the binding.

In addition to modifying how the binding process itself works, we make
one more change here.  Previously if we wanted to make changes to a C
method from the Python layer, we just replaced it by name and used
super() to access the C version.  mypy doesn't like that, since we're
effectively overriding a method with a differently-typed method.
Instead, we now prefix the names of these C methods with underscores and
call them by their underscore names from the superclass.

There are a few loose ends here — handling of callback types isn't
really beautiful, for example, but we always wrap that from the higher
layers, anyway, so we can leave it as is for the time being.