pyproject.toml: overhaul tox config, again
The new setup has some nice properties, mostly related to building
wheels in a separate environment:

 - pip doesn't try to create a virtual environment for installing from
   sdist, which means that we can always used the cached environment
   from tox.  That means that, after the first run for a given
   environment, the setup works offline.  It's also faster.

 - our default rules now flat-out refuse to download anything at all
   from PyPI and needs to have all dependencies available in the distro

 - because all venv-based target environments get their wheel built in
   the same environment, we only need to do the package build once
   during parallel (`tox p`) runs.

 - (perhaps most importantly) this enables us to test Python 3.6 without
   breaking flit or requiring an ancient version of setuptools.  The
   build environment is either the distro (pkg) or a venv with the
   default installed Python (venv-pkg).  The wheel build takes place
   there, and the only thing the pip in the target venv needs to support
   is installing the resulting wheel.

 - linters (mypy, pycodestyle, ruff) are invoked separately from pytest,
   which is helpful because (for example) mypy doesn't work with 3.6.

Slightly annoyingly, tox insists on separate environments for packaging
and testing, which means we can't use the site environment for building
itself.  Strictly speaking, that's correct, though: we shouldn't have
the packaging dependencies installed in the testing environment, and
vice versa.  It's also easy to work around with `base=`.

We have to adjust our workflows to reflect the fact that tox will never
download anything for itself anymore: this requires us to install some
more testing dependencies from the distro level (or in the case of RHEL,
from pip, ahead of time).  That's good: it exposes us to a wider range
of environments and validates the idea that our tox tests run against
the distro, not stuff from PyPI.

Unfortunately we can't enable Python 3.6 testing because the macro we
use to get the various Python versions no longer supports it.  It
works locally on Fedora 38, however.