Use hardened xml parser defusedxml
It took entirely too long, but currently released OS distributions now
have a safe by default ElementTree, but I think this may be used by some
people on older systems that are still vulnerable by shipping it via ssh
to that old system from a new system with cockpit.

The xml here comes from something that is in a normal configuration not
untrusted (a process running as a local user receives the xml from the
systems dbus/systemd instance) and it doesn't cross a possible security
boundary between cockpit-bridge and cockpit-ws, as this is probably only
used in cockpit-bridge. On the other hand this is published on pypi.

I considered if this is not needed, overall if this change just works,
it is probably still a good idea.