bustypes: add support for 'h' (fd passing)
Add support for passing file descriptors via D-Bus.  This is done via
Handle instances.

On the receiver side this is a nice way to make sure that the fd gets
properly closed when the user is done with it.

On the sending side it's a nice way to ensure that the user is
definitely intending to send a file descriptor.  This is important for
users like cockpit-bridge which might receive raw JSON containing
messages containing variants of type "h" and raw integers: if we just
pass those messages through (as cockpit-bridge does) then those integers
will be interpreted as fds of the cockpit-bridge process, which is
definitely not something we want.  Forcing the user to explicitly create
a Handle() instance guards against this.

Add some test cases, written by Claude.